+++ core/module/config/view/backup/backup.php
@@ -10,21 +10,24 @@
 	</div>
 	<div class="col2 offset8">
 		<?php echo template::submit('configBackupSubmit',[
-			'value' => 'Valider',
-			'ico' => 'check'
-			]); ?>
+			'value' => 'Sauvegarder'
+		]); ?>
 	</div>
+	<div class="modal">Merci de patienter, je travaille pour vous.<!-- Emplacement pour l'animation --></div>
 </div>
 <div class="row">
 	<div class="col12">
 		<div class="block">
 			<h4>Paramètre</h4>
 			<div class="row">
-				<div class="col8 offset1">
+				<div class="col12">
 					<?php echo template::checkbox('configBackupOption', true, 'Inclure le contenu du gestionnaire de fichiers', [
 						'checked' => true,
-						'help' => 'Cette option n\'est pas recommandée lorsque le contenu du gestionnaire de fichiers est très volumineux.'
+						'help' => 'Si le contenu du gestionnaire de fichiers est très volumineux, mieux vaut une copie par FTP.'
 					]); ?>
+				</div>
+				<div class="col12">
+					<em>L'archive est générée dans <a href="<?php echo helper::baseUrl(false); ?>core/vendor/filemanager/dialog.php?fldr=backup&type=0&akey=<?php echo md5_file(self::DATA_DIR.'core.json'); ?>"  data-lity>le dossier Backup</a> du gestionnaire de fichiers.</em>
 				</div>
 			</div>
 		</div>