+++ core/module/config/view/manage/manage.php
@@ -10,8 +10,7 @@
 	</div>
 	<div class="col2 offset8">
 		<?php echo template::submit('configManageSubmit',[
-			'value' => 'Valider',
-			'ico' => 'check'
+			'value' => 'Restaurer'
 		]); ?>
 	</div>
 </div>