+++ core/module/install/ressource/defaultdata.php
@@ -5,10 +5,13 @@ class init extends common {
 			'analyticsId' => '',
 			'autoBackup' => true,
 			'autoUpdate' => true,
+			'autoUpdateHtaccess' => false,
 			'cookieConsent' => true,
 			'favicon' => 'favicon.ico',
 			'faviconDark' => 'faviconDark.ico',
 			'homePageId' => 'accueil',
+			'page404' => 'erreur404',
+			'page403' => 'erreur403',
 			'maintenance' => false,
 			'social' => [
 				'facebookId' => 'facebook',
@@ -24,20 +27,20 @@ class init extends common {
 			'legalPageId' => 'mentions-legales',
 			'metaDescription' => 'Zwii est un CMS sans base de données qui permet de créer et gérer facilement un site web sans aucune connaissance en programmation.',
 			'title' => 'Votre site en quelques clics !',
-			'proxyUrl' => "",
-			'proxyPort' => "",
-			'proxyType' => "tcp://",
+			'proxyUrl' => '',
+			'proxyPort' => '',
+			'proxyType' => 'tcp://',
 			'smtp' => [
 				'enable' => false,
 			],
-			"connect" => [
-				"timeout" => 600,
-				"attempt" => 3,
+			'connect' => [
+				'timeout' => 600,
+				'attempt' => 3,
 				'log' => false
 			]
 		],
 		'core' => [
-			'dataVersion' => 10000,
+			'dataVersion' => 10300,
 			'lastBackup' => 0,
 			'lastClearTmp' => 0,
 			'lastAutoUpdate' => 0,
@@ -109,10 +112,11 @@ class init extends common {
 				'displayCopyright' => false,
 				'displayLegal' => false,
 				'displaySearch' => false,
+				'displayMemberBar' => false,
 				'template' => '3'
 			],
 			'header' => [
-				'backgroundColor' => 'rgba(255, 255, 255, 1)',
+				'backgroundColor' => 'rgba(32, 59, 82, 1)',
 				'font' => 'Oswald',
 				'fontSize' => '2em',
 				'fontWeight' => 'normal',
@@ -123,7 +127,7 @@ class init extends common {
 				'margin' => false,
 				'position' => 'site',
 				'textAlign' => 'center',
-				'textColor' => 'rgba(0, 17, 157, 1)',
+				'textColor' => 'rgba(255, 255, 255, 1)',
 				'textHide' => false,
 				'textTransform' => 'none',
 				'linkHomePage' => true,
@@ -146,7 +150,8 @@ class init extends common {
 				'activeColorAuto' => true,
 				'activeColor' => 'rgba(255, 255, 255, 1)',
 				'activeTextColor' => 'rgba(255, 255, 255, 1)',
-				'radius' => '0px'
+				'radius' => '0px',
+				'memberBar' => true
 			],
 			'site' => [
 				'backgroundColor' => 'rgba(255, 255, 255, 1)',
@@ -537,7 +542,53 @@ class init extends common {
 				'hideMenuSide' => false,
 				'hideMenuHead' => false,
 				'hideMenuChildren' => false
-			]
+			],
+			'erreur403' => [
+				'typeMenu' => 'text',
+				'iconUrl' => '',
+				'disable' => false,
+				'content' => '<h2 style="text-align: center;">Vous n\'êtes pas autorisé à accéder à cette page...</h2><p style="text-align: center;">Personnalisez cette page à votre convenance sans qu\'elle apparaisse dans les menus.<p>',
+				'hideTitle' => false,
+				'breadCrumb' => false,
+				'metaDescription' => '',
+				'metaTitle' => '',
+				'moduleId' => '',
+				'modulePosition' => 'bottom',
+				'parentPageId' => '',
+				'position' => 0,
+				'group' => self::GROUP_VISITOR,
+				'targetBlank' => false,
+				'title' => 'Erreur 403',
+				'block' => '12',
+				'barLeft' => '',
+				'barRight' => '',
+				'displayMenu' => 'none',
+				'hideMenuSide' => false,
+				'hideMenuChildren' =>false
+			],
+			'erreur404' => [
+				'typeMenu' => 'text',
+				'iconUrl' => '',
+				'disable' => false,
+				'content' => '<h2 style="text-align: center;">Oups ! La page demandée est introuvable...</h2><p style="text-align: center;">Personnalisez cette page à votre convenance sans qu\'elle apparaisse dans les menus.<p>',
+				'hideTitle' => false,
+				'breadCrumb' => false,
+				'metaDescription' => '',
+				'metaTitle' => '',
+				'moduleId' => '',
+				'modulePosition' => 'bottom',
+				'parentPageId' => '',
+				'position' => 0,
+				'group' => self::GROUP_VISITOR,
+				'targetBlank' => false,
+				'title' => 'Erreur 404',
+				'block' => '12',
+				'barLeft' => '',
+				'barRight' => '',
+				'displayMenu' => 'none',
+				'hideMenuSide' => false,
+				'hideMenuChildren' =>false
+			],
 		],
 		'module' => [
 			'blog' => [
@@ -632,7 +683,7 @@ class init extends common {
 			'contact' => [
 				'config' => [
 					'button' => '',
-					'capcha' => true,
+					'captcha' => true,
 					'group' => self::GROUP_ADMIN,
 					'pageId' => '',
 					'subject' => ''
@@ -665,3 +716,4 @@ class init extends common {
 		]
     ];
 }
+