+++ core/module/install/view/index/index.php
@@ -20,7 +20,7 @@
 	</div>
 	<?php echo template::mail('installMail', [
 		'autocomplete' => 'off',
-		'label' => 'Adresse mail'		
+		'label' => 'Adresse mail'
 	]); ?>
 	<div class="row">
 		<div class="col6">
@@ -41,7 +41,7 @@
 			<?php echo template::checkbox('installDefaultData',true , 'Charger un exemple de site', [
 				'checked' => true
 			]);
-			?> 
+			?>
 		</div>
 	</div>
 	<div class="row">